requirements.txt
keras==2.9.0
matplotlib==3.6.1
numpy==1.23.4
opencv_python==4.6.0.66
scikit_learn==1.1.3
tensorflow_gpu==2.9.1
torch==1.12.0+cu116
tqdm==4.64.1
